+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+**0.3.0**
+
+-   Replace equipment pickup model, icon, name and description with scrap and "???"
+-   Store EquipmentDefs and pass to notification queue
+
+**0.2.0**
+
+-   Store ItemDefs and pass to notification queue
+
+**0.1.1**
+
+-   Correct path to void scrap icons
+
 **0.1.1**
 
 -   Correct path to void scrap icons